requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
affine=2.3.0=py_0
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
blosc=1.21.0=h0e60522_0
boost-cpp=1.74.0=h5b4e17d_4
brotlipy=0.7.0=py39hb82d6ee_1001
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.7.5=haa95532_1
cairo=1.16.0=hb19e0ff_1008
certifi=2021.5.30=py39haa95532_0
cffi=1.14.6=py39h0878f49_0
cfitsio=3.470=h0af3d06_7
chardet=4.0.0=py39hcbf5309_1
charls=2.2.0=h39d44d4_0
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_0
coverage=5.5=py39hb82d6ee_0
cryptography=3.4.7=py39hd8d06c1_0
curl=7.77.0=h789b8ee_0
cycler=0.10.0=py_2
decorator=5.0.9=pyhd8ed1ab_0
draco=1.4.1=h5362a0b_0
exiv2=0.27.1=hfa2c1cf_0
expat=2.4.1=h39d44d4_0
fiona=1.8.20=py39hea8b339_1
fontconfig=2.13.1=h1989441_1005
freeglut=3.2.1=h0e60522_2
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
future=0.18.2=py39hcbf5309_3
gdal=3.3.1=py39h6795fcd_0
geopandas=0.9.0=pyhd8ed1ab_1
geopandas-base=0.9.0=pyhd8ed1ab_1
geos=3.9.1=h39d44d4_2
geotiff=1.6.0=ha8a8a2d_6
gettext=0.19.8.1=h1a89ca6_1005
giflib=5.2.1=h8d14728_2
gsl=2.7=hdfb1a43_0
hdf4=4.2.15=h0e5069d_3
hdf5=1.10.6=nompi_h5268f04_1114
httplib2=0.19.1=pyhd8ed1ab_0
icu=68.1=h0e60522_0
idna=3.1=pyhd3deb0d_0
imagecodecs=2021.7.30=py39h166567b_0
intel-openmp=2021.3.0=h57928b3_3372
jasper=2.0.14=h77af90b_2
jbig=2.1=h8d14728_2003
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jsoncpp=1.9.4=h2d74725_3
jxrlib=1.1=h8ffe710_2
kealib=1.4.14=h96bfa42_2
kiwisolver=1.3.1=py39h2e07f2f_1
krb5=1.19.1=hbae68bd_0
laszip=3.4.3=h6538335_1
laz-perf=2.1.0=py39h415ef7b_3
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libaec=1.0.5=h39d44d4_0
libblas=3.9.0=9_mkl
libbrotlicommon=1.0.9=h8ffe710_5
libbrotlidec=1.0.9=h8ffe710_5
libbrotlienc=1.0.9=h8ffe710_5
libcblas=3.9.0=9_mkl
libclang=11.1.0=default_h5c34c98_1
libcurl=7.77.0=h789b8ee_0
libdeflate=1.8=h8ffe710_0
libffi=3.3=h0e60522_2
libgdal=3.3.1=ha6cada3_0
libglib=2.68.3=h1e62bf3_0
libiconv=1.16=he774522_0
libkml=1.3.0=h9859afa_1013
liblapack=3.9.0=9_mkl
liblapacke=3.9.0=9_mkl
libnetcdf=4.8.0=nompi_hf689e7d_103
libopencv=4.5.2=py39h41a9e09_1
libpng=1.6.37=h1d00b33_2
libpq=13.3=hfcc5ef8_0
libprotobuf=3.16.0=h7755175_0
librttopo=1.1.0=hb340de5_6
libspatialindex=1.9.3=h39d44d4_3
libspatialite=5.0.1=h762a7f4_5
libssh2=1.9.0=h680486a_6
libtiff=4.3.0=h0c97f57_0
libwebp=1.2.0=h2bbff1b_0
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.12=hf5bbc77_0
libxslt=1.1.33=h65864e5_2
libzip=1.8.0=hfed4ece_0
libzopfli=1.0.3=h0e60522_0
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mapclassify=2.4.2=pyhd8ed1ab_0
markupsafe=2.0.1=py39hb82d6ee_0
matplotlib-base=3.4.2=py39h581301d_0
mkl=2021.2.0=hb70f87d_389
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.0=py39h277e83a_2
mkl_random=1.2.2=py39hf11a4ad_0
mock=4.0.3=py39hcbf5309_1
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
networkx=2.3=py_0
nitro=2.7.dev6=h33f27b4_3
nose2=0.9.2=py_0
numpy=1.20.1=py39h34a8a5c_0
numpy-base=1.20.1=py39haf7ebc8_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.2=py39hcbf5309_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1l=h2bbff1b_0
owslib=0.25.0=pyhd8ed1ab_0
pandas=1.3.0=py39h2e25243_0
pcre=8.45=h0e60522_0
pdal=2.3.0=hde8ebe7_3
pillow=8.3.1=py39h916092e_0
pip=21.1.3=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
pkg-config=0.29.2=h2bf4dc2_1008
pkgconfig=1.5.5=py39hcbf5309_0
plotly=5.3.1=pyhd8ed1ab_0
poppler=21.03.0=h9ff6ed8_0
poppler-data=0.4.10=0
postgresql=13.3=h1c22c4f_0
proj=8.0.1=h1cfcee9_0
psycopg2=2.9.1=py39h0878f49_0
py-opencv=4.5.2=py39h832f523_1
pycparser=2.20=pyh9f0ad1d_2
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=3.1.0=py39ha996c60_3
pyqt=5.12.3=py39hcbf5309_7
pyqt-impl=5.12.3=py39h415ef7b_7
pyqt5-sip=4.19.18=py39h415ef7b_7
pyqtchart=5.12=py39h415ef7b_7
pyqtwebengine=5.12.1=py39h415ef7b_7
pyqtwebkit=5.212=py39h8235778_1
pysocks=1.7.1=py39hcbf5309_3
python=3.9.6=h7840368_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py39hb82d6ee_1
qca=2.2.1=hd7ce7fb_3
qgis=3.18.3=py39h2c41444_2
qjson=0.9.0=hd7ce7fb_1006
qscintilla2=2.11.2=py39h8235778_4
qt=5.12.9=h5909a2a_4
qtkeychain=0.12.0=h552f0f6_0
qtlocation=5.12.9=ha925a31_0
qtwebkit=5.212=hb258684_1
qwt=6.1.6=h552f0f6_0
qwtpolar=1.1.1=hd7ce7fb_7
rasterio=1.2.6=py39h85efae1_2
requests=2.26.0=pyhd8ed1ab_0
rtree=0.9.7=py39h09fdee3_1
scikit-learn=0.24.2=py39he931e04_0
scipy=1.7.0=py39hc0c34ad_0
setuptools=49.6.0=py39hcbf5309_3
shapely=1.7.1=py39haadaec5_5
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=ha925a31_3
snuggs=1.4.7=py_0
sqlite=3.36.0=h8ffe710_0
tbb=2021.3.0=h2d74725_0
tenacity=8.0.1=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh8a188c0_0
tifffile=2021.8.8=pyhd8ed1ab_0
tiledb=2.3.2=h78dabda_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py39hb82d6ee_1
tzdata=2021a=he74cb21_1
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wheel=0.36.2=pyhd3deb0d_0
win_inet_pton=1.1.0=py39hcbf5309_2
wincertstore=0.2=py39hcbf5309_1006
xerces-c=3.2.3=h0e60522_2
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zfp=0.5.5=h0e60522_5
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
